# apis/openapi.yml
openapi: 3.0.3
info:
  title: execute_academy API
  description: REST API for execute_academy. See category, product, auth, and user endpoints.
  license:
    name: ''
  version: 0.1.0
servers:
- url: http://localhost:9098
  description: Local dev
paths:
...
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: Register a new user account
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error or bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/crate.pkg.response.ApiErrorResponse'
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/crate.pkg.response.ApiErrorResponse'
...
  /api/lessons/:id:
    delete:
      tags:
      - Courses
      operationId: delete_lesson
      responses:
        '200':
          description: Deleted
    patch:
      tags:
      - Courses
      operationId: update_lesson
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLessonRequest'
        required: true
      responses:
        '200':
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
...
components:
  schemas:
...
    Lesson:
      type: object
      required:
      - id
      - module_id
      - title
      - duration
      - position
      - is_free
      - published
      - created_at
      properties:
        content:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        description:
          type: string
          nullable: true
        duration:
          type: string
        id:
          type: string
          format: uuid
        is_free:
          type: boolean
        module_id:
          type: string
          format: uuid
        position:
          type: integer
          format: int32
        published:
          type: boolean
        title:
          type: string
        updated_at:
          type: string
          format: date-time
          nullable: true
        video_url:
          type: string
          nullable: true
...
    RegisterRequest:
      type: object
      description: |-
        Common HTTP request/response types for the Auth and Users routes.

        These are intentionally separated from the route handlers to reduce
        duplication and keep a single source of truth for our API schemas.
      required:
      - first_name
      - last_name
      - username
      - email
      - password
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
          maxLength: 50
          minLength: 1
        last_name:
          type: string
          maxLength: 50
          minLength: 1
        password:
          type: string
          minLength: 8
        username:
          type: string
          maxLength: 30
          minLength: 3
          pattern: ^[a-zA-Z0-9_]+$
    RegisterResponse:
      type: object
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
...
    UpdateLessonRequest:
      type: object
      properties:
        content:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
        is_free:
          type: boolean
          nullable: true
        position:
          type: integer
          format: int32
          nullable: true
        published:
          type: boolean
          nullable: true
        title:
          type: string
          nullable: true
        video_url:
          type: string
          nullable: true
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT